ipipe: Introduce infrastructure for userspace return notifier
A little bit inspired by the kernel's user return notifier, this
introduces an I-pipe hook before the kernel jumps back to a userspace
context from the root domain. The hook is design to allow a switch back
to the head domain, thus will not run through signal/preemption checks
when returning from the callback over head.

The first use case for the hook is resumption of ptraced tasks over
head if they were stopped in that domain.