x86/ipipe: Disable access_ok context under I-pipe
The test in_task() does not work in the head domain and can trigger a
false positive warning when executed over a real-time task. As this
header is too central, we cannot pull in pipe.h in order to gain
access to ipipe_root_p and confine the check. So disable it for I-pipe.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>